Rename Statamic\Ignition namespace to Statamic\ErrorSolutions
