+++ src/Exceptions/ComposerJsonMissingPreUpdateCmdException.php
@@ -5,7 +5,7 @@ namespace Statamic\Exceptions;
 use Exception;
 use Spatie\ErrorSolutions\Contracts\ProvidesSolution;
 use Spatie\ErrorSolutions\Contracts\Solution;
-use Statamic\Ignition\Solutions\EnableComposerUpdateScripts;
+use Statamic\ErrorSolutions\Solutions\EnableComposerUpdateScripts;
 
 class ComposerJsonMissingPreUpdateCmdException extends Exception implements ProvidesSolution
 {
