+++ src/Exceptions/StatamicProRequiredException.php
@@ -5,7 +5,7 @@ namespace Statamic\Exceptions;
 use Exception;
 use Spatie\ErrorSolutions\Contracts\ProvidesSolution;
 use Spatie\ErrorSolutions\Contracts\Solution;
-use Statamic\Ignition\Solutions\EnableStatamicPro;
+use Statamic\ErrorSolutions\Solutions\EnableStatamicPro;
 
 class StatamicProRequiredException extends Exception implements ProvidesSolution
 {
