+++ src/Providers/IgnitionServiceProvider.php
@@ -5,8 +5,8 @@ namespace Statamic\Providers;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Support\ServiceProvider;
 use Spatie\ErrorSolutions\Contracts\SolutionProviderRepository;
-use Statamic\Ignition\SolutionProviders\OAuthDisabled;
-use Statamic\Ignition\SolutionProviders\UsingOldClass;
+use Statamic\ErrorSolutions\SolutionProviders\OAuthDisabled;
+use Statamic\ErrorSolutions\SolutionProviders\UsingOldClass;
 
 class IgnitionServiceProvider extends ServiceProvider
 {
